load continent into flecs

+++ core/src/main/java/com/populaire/projetguerrefroide/component/Ideology.java
@@ -3,5 +3,5 @@ package com.populaire.projetguerrefroide.component;
 import com.github.elebras1.flecs.annotation.FlecsComponent;
 
 @FlecsComponent
-public record Ideology(int color, short factionDriftingSpeed) {
+public record Ideology(int color, byte factionDriftingSpeed) {
 }
